packages/pantheon-core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pantheon-debate"
version = "0.1.0a0"
description = "A multi-agent debate framework — summon gods, prophets, and historical figures to deliberate on any question."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Wendy" }]
keywords = ["llm", "multi-agent", "debate", "mcp", "openclaw"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "typer>=0.12",
    "rich>=13.7",
]

[project.optional-dependencies]
classifier = ["sentence-transformers>=2.7"]
otel = ["opentelemetry-api>=1.24", "opentelemetry-sdk>=1.24"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4", "mypy>=1.10"]

[project.scripts]
pantheon = "pantheon.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/pantheon"]

[tool.hatch.build.targets.wheel.shared-data]
"personas" = "pantheon/personas/_builtin"
